problem 1 marked

+++ cpp/ps0001/ps0001_slon.cpp
@@ -2,15 +2,19 @@
 using namespace std;
 int main()
 {
-    int a;
-    int b;
-    int c;
+    // promptiung user for 3 numbers to add 1/2
+    short a;
+    short b;
+    short c;
     cin >> a; "\n";
     cin >> b; "\n";
     cin >> c; "\n";
+    
+    // outputting the result of the summation process 3/3
     cout <<"============================\n";
     cout <<"Answer is   ";
     cout << a + b + c;
 
     return 0;
 }
+// great work 😎 8/10 ⭐
